From: "Luis F. Ortiz" <lfo@polyad.org> via: Rene Mayrhofer <rmayr@debian.org>
The structure member pcr_len is a size_t (long) and is the first member
of the structure.  The address 0xeff34bae is NOT kosher on a SPARC for a
long; it needs to be aligned on a long word boundry to not cause
alignment problems.  This problem happens all the time when people port
stuff from x86 to SPARC machines.  GDB will read it fine, but the
program will fault trying to use the pointer.

One possible solution, that should be safe for all architectures, is to
force the alignment of the 'reqbuf' to be the same as the alignment of
the first structure member.  Since this structure is only used for
communication within the same machine, no other alignment issues will
arise.

Signed-Off-By: lfo@polyad.org