From: "Matthias Haas" <mh@pompase.net> Date: Thu, 18 May 2006 16:41:52 +0200 (CEST)
I adapted Mathieu Lafons (at open-source.arkoon.net) patch to openswan 1.0
for openswan 2.4.5, to allow rsa and pak authentication for roadwarriors
(with dyn ips). This patch is at a level to say it works for me :-). any
further tests and comments are welcome.
This patch does not resolve the limitation that all roadwarriors share the
same psk, but you can have roadwarrios that do rsa based auth and
roadwarriors using psk.

Signed-off-by: Michael Richardson <mcr@xelerance.com>