Under situations of load, where many requests can be queued at the same time, due the way that pipes and stdio buffers, the helper can wind up confusing the stdio buffers, and wind up reading back it's own output. The result is a mess.
stdio says that you can't mix reads/writes unless you do some kind of
file positioning operation to synchronize state. Rather than doing that
on a pipe, let's just use an input and an output stdio buffer.

The reason we are using stdio at all is so that we can let it worry about
all the buffering, and reading from the kernel in bigger chunks.

In addition the above, the parent pluto process could find itself in
a short write, so centralize the recover of that.