Fix for bug #460 by Herbert Xu. From the bug entry:
When main mode I2->R2 is delayed for crypto, it may pick up a bogus
ISAKMP header when resuming. This is because reply_buffer is global
and another transaction could've used it in the time being.

So we simply generate the header in main_inI2_outR2_tail instead of
process_packet for this case.