Here's the verbose description since I think this can be cleaned up, I am just not prepared to do it right now.
The previous %any fix created a problem with default ipsec.secret entries,
that is the ones with no peers specified:

	: PSK "..."

It used to work because two ID_NONE's are added to secret entries with no ID's.
Because same_id always matches ID_NONE,  we ended up with a
match_him|match_me answer.  Which IMO is wrong since it can trump a true
match of him/me.

As any_id now also matches ID_NONE, the %any change meant we ended up with a
"match_any" result that was not handled at all.

The solution for now,  add in a "match_any" case,  which is mostly good
as it allows specific matches to be chosen in preference, and it also means
that slightly better matches (match_him|match_any) will also be chosen
above the complete wildcard match.