When a compressed packed is received, the kernel panics (also see http://bugs.xelerance.com/view.php?id=982 , which appears to describe the same issue).
...

We observe this with 2.4.13, but this appears to be present
in 2.6.19 as well.

It is due to bogus offset calculations in ipcomp.c:skb_copy_ipcomp().
"offset" is the difference between two skbs, yet the value is used to
set the skb protocol header fields.

offset=n->head-skb->head;
n->nh.raw=skb->nh.raw+offset;

got changed to

offset=n->head-skb->head;
skb_set_network_header(n, offset);

which will make the network header point to somewhere
outside the skbs allocated memory area,
causing the kernel to OOOPS on the first receipt of a compressed packet.

This makes things work again for us with 2.4.13 on a Linux 2.6.16-based kernel,
it would be great if you could double-check this because we don't use all
the code paths there due to #ifdefs.

Florian Westphal <fwestphal@astaro.com>