Implement a fallback to SW for failed HW requests.
Both the IXP and the safenet fail with "out of range" like errors
when using RSA keys of 1024 bits (and possibly 512 bits).

Implement a system whereby, if mod_exp or mod_exp_crt fails,  we fall
back to the software version.

This is a first pass and implements enough for us to do more things later.
It's possible we should give up if the failure rates are too high when using
HW.