A bunch of little fixups for linux-2.6.29.
For the most part it is cleaning up the use of the "priv" field of the
struct net_device. It is no longer a field in its own right, and you
need to use the netdev_priv() accessor function to get at it. (Along
with this you must also allocate it as part of the alloc_netdev() call).

Also had to convert the use of the task_struct's uid field into a
capability check.