Remove duplicate virtual interface assigning code.
Do not do the virtual interface assigning in the NAT-T code as it is already
being done in ipsec_rsm for all receive packets.

The original problem that this code fixed could have also be fixed by
removing the default assigned of "ipsec0" to all packets, this change
effectively does that.

Since the NAT-T version had some useful debug, pull some of that into the
ipsec_rsm version of the code to help with debugging this kind of thing
in the future.