Patch by Anthony Tong <atong@TrustedCS.com>
We found this is 2.4 when tracking down an issue why sometimes concurrent
ISAKMP negotiations from different hosts to a single host would fail
(nhelpers>=1). On sending the main r2 message, there is a possibility that
its message header belongs to another ike message that is originally
intended for another host. (the correct one was overwritten due to
continuation)