Compile fix by D. Hugh Redelmeier for this error on OSX:  Undefined symbols:   "_oswcrypto", referenced from:     _oswcrypto$non_lazy_ptr in libopenswan.a(pem.o)  ld: symbol(s) not found
His explanation:

C had declarations of objects.  A subset of them define objects. A
declaration says: this is the type of this thing and a few other details
like its storage class.  But it doesn't make it exist.
A definition makes the thing exist.
Strict C allows only one definition in all the files that make up a program.
Something like "extern int i;" is surely a declaration that isn't a definition.
Something like "int i = 3;" is surely a definition.
Something like "int i;" is a tentative definition.  It works as a definition
unless something better comes along.  (roughly)
So the change turned a tentative definition into a for-sure definition.
It should not be needed. I don't know why this build system (xcode?) is
requiring it.