Fix up the "patchless" NAT-T support in klips
This gets klips and nat-t working.  The decapsulation code is still needed
because we cannot use the code path in the kernel that net_key uses,  so
the klips version is just a dupe of the net_key/xfrm kernel code that isn't
there if you are using klips ;-).

I also cleaned up all the differing NAT_TRAVERSAL ifdefs to be common as
CONFIG_IPSEC_NAT_TRAVERSAL will not longer exist.

I made pluto auto detect the best NAT-T approach for the klips stack.

I have tested this with linux-2.6.29, with and without the old nat-t patch
using both klips and netkey.  All four combos worked for my tests :-)
I haven't tested with an old pre 2.6.23 kernel.

Special thanks to Harald Jenny whose original klips patch was the basis
for much of this implementation.