A patch for an iops due to improper ipsec_sa destruction. This is bug #1023 See https://bugs.xelerance.com/view.php?id=1023 Contributed by Nick Jones <nick.jones@network-box.com>
From nick:
There is a discrepancy in the way that ipsec_sa objects are initialised
and destroyed, causing a kernel Oops in ipsec.ko while destroying an
ipsec_sa that was found to be invalid part way through initialisation.

This happens when adding an spi with the same id as one already established,
but we are also seeing this happen at runtime on heavily loaded boxes.

The solution I am suggesting is to perform a more mirrored init/destroy
by tying more closely the ips_key_e and ips_alg_enc fields of the
ipsec_sa structure, similarly the same should be done for ips_key_a
and ips_alg_auth. This patch does this by removing the call to
ipsec_alg_sa_init in pfkey_sa_process and moving it (actually the
logic therein) to ipsec_alg_enc_key_create, so that the initialised
key structure is guaranteed to match the algorithm used to create (and
destroy) it.