Partially started opportunistic init breaks phase2
Do not start any opportunistic operations unless we belong to an opportunistic
tunnel.  Installing the bare shunt on a non-opportunistic tunnel causes us
to fail to "replace" the eroute later as there is actually no eroute and
we should be doing an "add" (we do a replace because bspp is non null in
pluto's route_and_eroute function).

This problem shows on "routed" tunnels that are not up (ie., auto=route).
If you ping an address on the remote network,  openswan will start an
opportunistic neg. as above, abandon it as the tunnel is not opportunistic,
but leave the bare shunt in place.

The symptom after this happens is a tunnel the doesn't ever complete phase2
and gets stuck in an I1/R1 state start rather than complete and make it to
a I2/R2 state.

There should be a K_SADB_X_ADDFLOW error in your syslog,
errno -14, Bad address (EFAULT).