Partial bugfix for #1004. We now get the right protoport settings in ip xfrm pol.
The problem was that the ports are defined in two places, one is port
and another is within an ip_struct. The old whack code ran an
update_ports() on the whack message to synchronise these two different
locations. This code was missing in the new starterwhack.c code.

Patch by DHR and Paul.