Cleanup up narrow hold eroutes when tunnel is up
Currently there is a problem where traffic on routed road warrior tunnels
produces narrow hold eroutes that are not managed.

It seems there may also be some race condiitons causing hold eroute to
get missed.

These problems can prevent tunnels reaching their final states,  or just
prevent traffic between hosts over the tunnel.

The previous commits were reverted as they were incorrectly stopping hold
routes on non-opportunistic tunnels:

	6dcb7f540005bdb2894736c59048a2812499ab71
	6e8b4648b735c9b222f0784d6894b4f0b72f85a3

Here we opt to clean up at a point when we know things should be good.
This is less than ideal but the interactions and dependencies within the
shunt/eroute code are tricky to follow and change without regression, so
the safe approach for now is to fix it up when we know what should be
happening.

As part of this,  we also get to clean up what may be a large number of
host-host hold eroutes,  that previously would have been missed.