Prevent aggressive mode tunnels losing phase2
Make sure we update any pending SA requirements with the new SA when we
replace it. Without this we end up with a phase2'less aggressive mode tunnel
that thinks everything is ok, phase2 will never come online.

The easiest way to create such a phase2'less tunnel is:

    * central server running openswan with an Agressive mode road-warrior
      connection.

    * connect in with the road warrior client (DPD settings of 9 and 30)

    * disable the tunnel at the server end in a way that sends a "Delete SA"
      to the RW for both the P1 and P2 SA's.  Something like:
      'whack --delete --name XXXX' will do it.

    * wait 2 or 3 minutes (gives RW phase1 and phase2 time to bail go back
      to initial state).

    * restart/enable the tunnel at the server end and wait for it to come up.

I am sure there there are some other combos that would show this behaviour,
but this was the easiest to reproduce that I found.

The problem is that when cleaning up the old SA it was still marked as
primary SA,  and so pending information for the connection was cleaned
as well,  thus removing the only thing that would kick in phase2 for us.