Make pluto respond with NAT-D info in AGGR mode
NAT traversal is aborting at the remote client when pluto responds to their
initiated connection (ie., road warrior).

The reason for this is that pluto does not send any NAT-D information when in
aggressive mode, so the remote client has nothing to work with.  The remote
client will see log messages like:

    NAT-Traversal: Only 0 NAT-D - Aborting NAT-Traversal negotiation

Which should not happen between two openswan systems even if the connection
is not NAT'd.

Also the debug for NAT traversal when using aggressive mode was somewhat
less that main mode, so bring it more into line with the main mode version
and iensure that interface information is accurate and working.