OCF fails to initialise IV on ESP packets
ESP packets sent using OCF acceleration have pseudo predictable IV's. The
OCF code fails to initialise the IV and thus gets portions of uninitialised
memory (quite often packet contents of IP addresses).

The possibility of predicting the IV's is not such a big problem
since the IV is visible in the resulting on-wire packets anyway.

A bigger issue is that this leaks 8 bytes of possibly sensitive
information in ESP packets that are traveling through unsafe networks.