Stop the queue when we are overloaded
Fix up klips to play nicer with the stack by stopping the queue
(return NETDEV_TX_BUSY) when we can no longer take any packets
and restarting it once we can.

This helps a lot with flow control through the stack back down to the NIC's
in high throughput scenarioes.