Fix oops when driver doesnt support all header ops
Handle the case where the device below us does not support all header ops.
This needs a better solution IMO, but this will hopefully get us by for now.

I think it might be better for us to report the same ops as the device below us.
At least until we have a better imp.