ipsecX route metric fix for Ubuntu 10.04
on startup, ipsecX route would have a metric of 0, but the systems' route for
the real device would have a route with metric 1.  That means, that as soon as
an SA with that route was initailzied, no packets woudl ever make it out on
that route.

This commit will detect a case where another route exists with a non-zero
metric that matches the spec assigned to the ipsecX device.  If that route
is found, it will be moved to a higher metric an the other routes.

At this time only Ubuntu 10.04 is known to create system routes with metric
of 1.  Other systems should not be effected by this change.

Signed-off-by: Bart Trojanowski <bart@jukie.net>