Changes to _realsetup.in for making the init script LSB compliant.
Currently, if pluto is killed (SIGKILL), it leaves both pid and lock
files, and if one does "service ipsec status; echo $?", it returns 2,
however for to be LSB compliant, it should return 1. The reason is that
when both files are left, then the script should check for pid file first,
and then for the lock file not the other way around. (rhbz #594767).

Signed-off-by: Paul Wouters <paul@xelerance.com>