convert allocation of saref subtables to kmalloc/GFP_ATOMIC
previously we used vmalloc() which resulted in a call to vmalloc() under
spin_lock_bh() after 2k sarefs have been given out.  vmalloc() calls BUG()
when called with preemption disabled -- which is bad.

It should be OK to allocate a 16k buffer (or 32k on 64bit systems) for the
subtable using kmalloc(), unless the system is under severe memory pressure.
Either way it's nicer to fail with a memory allocation failure than to panic
the kernel.