ip_select_ident needs to be called before hashing
I found,  by chance, the AH xmit path for klips protocol stack is a bit
broken. After a small research I found the ident field for ip header is
selected after generation of a hash for a packet. According to RFC 2402
ident must be selected before generation of a hash.

A possible fix is in the attachment, I hope it will be usefull.

Kirill.