First IPv6 packets through klips
This is a first rough pass over klips to add IPv6 support.
These changes allow an IPv6 over IPv4 tunnel to carry ping packets.
So a tunnel setup like this is possible:

fec0::1:0:0:0:1/128===10.31.1.9[+S=C]...10.31.1.3[+S=C]===fec0::2:0:0:0:1/128

There's lots of cleaning and rearranging to do and of course the IPv6 over IPv6
support needs to be done.  There are also lots of "not yet IPv6'd" code left
in there as well for good measure.