set ixs->outgoing_said in mast mode
Some code, specifically NAT-T, uses fields from the ixs->outgoing_said.  We
previously never set this structure in mast mode, because it was not needed
for looking up the SA, which is the case for protostack=klips.

Also added a mast_mode flag which is used to check if the xmit state is
working from a mast device or not.  This was previously done by checking
ixs->outgoing_said value for zero, which is no longer possible.