account for udp encap_rcv() function swap, and restore on unload
Previously when klips was told by pluto to convert a UDP socket to en
encapsulating one via the IPSEC_UDP_ENCAP_CONVERT ioctl, it would do
so blindly and w/o a way to restore the operation.  The conversion
replaced the 'encap_type' and 'encap_rcv' values of the socket, the latter
of which is a function call.

It's however possible that the socket would not be cleaned up until after
the klips module was unloaded.  This would mean that any packets arriving
after klips was unloaded, but before the socket was fully closed, would
end up calling the socket's 'encap_rcv' function pointer.  Since, at this
point in time, the pointer is garbage it would result in an Oops.

This fix models itself on the way that drivers/net/pppol2tp.c handles the
problem.