SAREF: rework exported interface to avoid might_sleep() calls during rcu lock
This rewrite avoids panic caused by...

    BUG: sleeping function called from invalid context at include/linux/kernel.h:157
    in_atomic(): 1, irqs_disabled(): 0, pid: 9671, name: xl2tpd
    Pid: 9671, comm: xl2tpd Not tainted 2.6.33.6-147.saref.fc13.x86_64 #1
    Call Trace:
    [<ffffffff8103a487>] __might_sleep+0xe8/0xea
    [<ffffffff813845ab>] might_fault+0x17/0x1e
    [<ffffffff81384625>] put_cmsg+0x73/0xe0
    [<ffffffffa023cc08>] klips_ip_cmsg_recv_ipsec_refinfo+0x7b/0xa4 [ipsec]
    [<ffffffff813bb46f>] ip_cmsg_recv+0x271/0x297
    [<ffffffff813d6f42>] udp_recvmsg+0x18d/0x217

The problem was caused by calling put_cmsg() while rcu lock was held in
ip_cmsg_recv_ipsec_refinfo().  The rcu lock is not intrussive, but it does
disable preemption, which is bad for put_cmsg() that needs to call
copy_to_user(), and could cause a page fault.

In this patch, I've moved the call to put_cmsg() to outside the rcu locked region.
It is thus no longer neccessary to pass the cmsg to klips for update, instead the
API was changed to get_secpath_sarefs() where klips returns the refme and refhim
values.

Similar change was made to the send path also.  API function was reduced to just
setting the SAref, no longer needs to know where it came from.