MAST: fix iptables rule "leak" on rekey
when an SA was rekeyed, we would try to delete the iptables rule matching
the nfmark generated from the new ref value of th SA that replaced it.
When coupled with an overlap window (configured with rekeymargin less
than salifetime) we would never actually remove the replcaed SAs in the
iptables chain... and the chain woudl grow indefinately.