MAST: use only the most recent iptables rule
iptables rules are evaluated in chain-order.  Since we add new rules to the top
of the chain, the most active rules are at the top and hit first.  However,
the chain evaluation didn't stop on the first hit.

This change makes the test expression of the chain more explicit, and will
only match packets that have not yet been marked.  This make sure that
we pick up the latest SA, in case junk is left behind in the table for
some reason.  It also means that counts in the iptables chain will be
correct.