MAST: fix how kernel_mast.c handles replacement of SAs with the same SA
This would occasionally happen: mast_sag_eroute() would be called to
replace an SA with the same one.  The mast_sag_eroute_replace() tries
to find the old SA and delete it, but since it was the one that was
just added, it would actually look like one was never added.