HAVESTATSD: Log new phase2 messages as a result of a rekey
The refresh_log code kept a tuple of tunnel/phase1/phase2 to surpress
duplicate messages, and considered a rekey of phase2 a duplicate. However,
a new SAref is generated, and this needs to go out via an HAVESTATSD
message. I now add the SAref to the LOG_CONN_STATSVAL(&lc) result. This
causes a new phase2 with saref to be different and a message to go out.

(David should definitely review this)