Revert "[NETKEY] Delete addresses from interface on down"
This reverts commit 411607ff44d5051da09ea021e7bbcb202feb995f.

It is too dangerous to delete any IP address if we don't keep a
list ourselves of what we added. We might be removing a core IP
address.

One possible solution is to add the IP to the loopback interface,
so that we better know what we can delete later on.