a better workaround for rp_filter
rp_filter follows the guidelines of RFC 1812 and incoming drops packets that
don't seem to make sense.  The particular test that KLIPS, in mast mode,
had problems with dropped packets that arrived on an interface to which
a reply packet would not be routed to.  For example, packet arrives on mast0,
if we reverse the src and dst addresses, we shold route that packet back
through mast0.  Because of the iptables + policy routing tricks mast
plays, we don't get that.

This commit does the foolwing:
 - packets arriving from a tunnel, and thus having skb->dev == mast0, will
   have the nfmark set to the SA they arrived on, along with the top bit set.

 - updown script now installs two ip-rules:
     from all iif mast0 lookup main
     from all fwmark 0x80000000/0x80000000 lookup 50
   the first one is new, and it sends all packets that arrive from a tunnel
   through the main table, and as before the nfmarked packets go through
   table 50.

 - rp_filter will reset the skb->nfmark unless we tell it that the interface
   set that mark.  We do this by setting the src_valid_mark to something
   arbitrary.  In our case we set it to 0x80000000.

All this to make rp_filter happy once more.