NETKEY: Fix for spurious kernel acquires landing us wrongly into         opportunistic code. [paul/dhr]
NETKEY is sending bogus aquire messages sometimes when used with
transport mode and L2TP (protoport 17/1701). These seem mostly triggered
by Windows and OSX clients. The acquire is notify of a %hold message in
the kernel. pluto then failed to match the acquire to a connection (there
was no matching on-demand tunnel or opportunistic connection) and would
install a "failsafe" %pass eroute. Unfortunately, the transport_proto
argument (17) was not set in this replacement eroute, and so we would
end up with a (bogus) %pass eroute, and a non-deleted netlink-aquire
%hold invisible to pluto.

In cannot_oppo() where we detect this failure, we now properly call
replace_bare_shunt() with the "delete" argument for the %hold.

A similar transport_proto mismatch in clear_narrow_holds() is fixed too.