| In programs/pluto/ike_alg.c at line 118, errbuf is assigned to errp. It is an | auto variable with a scope for just that function. Any use if that pointer | will be invalid for any callers.
Yes.

And the snprintf call seems to reflect a misunderstanding of the
length parameter.

And the return_on macro has flaws: references to parameters are not
protected with parens; the ugly do while idiom is used, but with a
semicolon at the end, defeating the purpose.  In each invocation, the
first argument is "ret", so it would be clearer and simpler to wire it
in.  In fact, the macro seems hardly worth the bother.

The buffer needs to be allocated by the caller OR use the heap.
There are two callers.

One (spdb_v1_struct.c line 1142) ignores the returned value
(immediately overwriting it).  So it should pass in NULL so that the
value isn't returned.  (Perhaps it should not ignore the returned
value.)

I've allocated the buffer as an auto in the caller and passed it into
ike_alg_enc_ok.