Fix to compile without DEBUG (bug #1040)
Note: sometimes there is a reliance on NO_DEBUG, so the safe way to actually
disable DEBUG is to change -DDEBUG to -DNO_DEBUG.

Mostly fixes to DBG macros, and moving variables inside the DBG macros.

Since this touched so many files, I also updated the Copyrights on these
as based on openswan-2 git history.