Bug #1005 Incorrect message "R_U_THERE_ACK has unexpected sequence number"
Description

There are several issues in the code that logs this message:

if (!p1st->st_dpd_expectseqno && seqno != p1st->st_dpd_expectseqno) {

     loglog(RC_LOG_SERIOUS, "R_U_THERE_ACK has unexpected sequence number (expected: %u got: %u", seqno, p1st->st_dpd_expectseqno);

"expected" and "got" are actually swapped, which makes this message
misleading. Also this message is triggered only when st_dpd_expectseqno
is zero, which as far as I understand means that the ACK response is
not expected at all

I have seen this message when the peer was slow and hasn't acknowledged
R_U_THERE in time. This issue may be related to 0000996

Signed-off-by: Paul Wouters <paul@xelerance.com>