Remove dependancy on atof() which does not exist on all embedded libc's
The ipsec.conf parser actually supported floats, but we have no keywords
that can take floats. Via ascii -> float it got set to 0 anyway, instead
of producing an error (for say keyingtries=4.5). This resolves an issue
on an arm cross compile where the libc implementation des not have atof()

Not yet fixed is the use of int instead of long for integer values. Since
all values are converted internally to seconds, this means we cannot
have more then a salifetime/lifetime of 0.3 days (32768 seconds)