Quickfix for /proc/net/ipsec/version to be world readable.
This is strange. The mode we set for this file in the proc_dir_entry
struct for "version" is (S_IFREG | S_IRUGO). We call all files to be
created with create_proc_entry() which also uses a mode to pass along,
which was always 0400. Perhaps the idea is that the kernel initialises
it to the mode passed, but then uses the mode from the struct later on?

However, that does not seem to happen. As far as I an tell, proc_dir_entry->mode
is completely ignored.

This quickfix changes the mode for our create_proc_entry() entry when
the name passed to it as arg1 is "version". I don't like this workaround.