Make ipsec addconn --configsetup more robust when broken conns are defined.
ipsec addconn would read the entire configuration file, regardless of how
it was called. When called with the --configsetup option, it should only
read the "conn setup" and return. To do this, confread_load() now takes an
additional bool parameter specifying if this is a "configsetup only" call.

Apart from optimising the call, I found when looking at another bug
that caused addconn to segfaul reading a regular conn, that not only the
conn, but the entire "config setup" section options were ignored. This
was caused by the conn breaking "ipsec addconn --configsetup". One bad
conn could therefor cause the entire "config setup" section to be ignored.
(I noticed when the listen= option failing to take effect)