Fix for https://bugs.openswan.org/issues/1095
When local esp or ah packets are marked with iptables like:

Chain OUTPUT (policy ACCEPT 5080 packets, 958K bytes)
 pkts bytes target     prot opt in     out     source               destination
 5080  958K MARK       all  --  *      *       0.0.0.0/0            0.0.0.0/0           MARK or 0x1

packets will rerouted due to the change of the mark in the OUTPUT
chain. The packet appears again on the ipsec device and will be dropped
with

klips_debug:ipsec_xmit_encap_bundle: shunt SA of DROP or no eroute:
dropping.

I think there must be also an exception for local esp and ah packets
in ipsec_tunnel_SAlookup (see patch) as for udp/500 and udp/4500. The
problem concerns both versions 2.4 and 2.6.

This was tested on kernel 2.6.22.19 and openswan 2.6.24 and openswan
2.4.15.