Cleanup use of IPSEC_NUM_IF and IPSEC_NUM_IFMAX
This was a bit of a mess with IPSEC_NUM_IF used when IPSEC_NUM_IFMAX was
intended in a number of places.  Also a couple of bad edge cases where
arrays could be indexed at IPSEC_NUM_IFMAX which is off the end.

Get all the usage consistent and fix the bugs.

As a side affect,  you can now configure ipsec4 with tncfg without having
ipsec3 configured,  which seemed to be the intent,  but was not possible.

/proc/net/ipsec_tncfg shows all configured devices now and not just those up
to IPSEC_NUM_IF.  Which is traditionally less than IPSEC_NUM_IFMAX.