KLIPS: Better interface handling in _startklips
Some code assumed ipsecX already existed. This might not be true anymore
in the near future (and was never true for ipsecX > 2)

We now tncfg --delete mast0 if we use protostack=klips
We now tncfg --delete ipsecX if we use protostack=mast

This should reduce clutter of unused interfaces.

Note that currently, ipsec0 cannot be deleted. This will get fixed soon.
Perhaps the code needs at least one of ipsecX or mastX for safety reasons
on handling /proc/net/ipsec/ files. If so, the module init code will have
to ensure we always have one type of virtual interface set.