osw_alias_cmp has a bug.
                s += nlen;
                while(*s!='\0' && *s!=' ' && *s!='\t') s++;

Why?  We should advance in s to the next possible match.  That is not
nlen characters hence, but only 1 character hence.  And then we should
start the attempt AFTER the next whitespace character, not at it.

So this code has probably never found a match that didn't start at
offset 0 in haystack.  Why?  Because every search after the first
starts looking at a whitespace character and that cannot match.

But the actuall bug that caused the dump is that the remainder of haystack might not
even be nlen long, so scan of s may skip the NUL at the end.

Does that mean that the scan is redundant?