Revert "If pluto is started with --nofork, then also disable nhelpers"
As Avesh pointed out, we always start with --nofork, because we want
the plutorun wrapper script to restart pluto on crash for us. I
changed this code so if HAVE_NO_FORK=true, then we disable fork and
nhelpers - but we no longer disable nhelpers when we saw --nofork.

This reverts commit 16989a3d0849ae8bb71df396d7f76ab4d6a63c03.

Conflicts:

	programs/pluto/plutomain.c