Openswan (IKEv2/IKEv1) icmp issue (redhat bz 681974):
It seems that Openswan does not have explicit support for processing
icmp traffic as specified in RFCs 4301/5996. Although IKEv1 (RFC 2409)
does not state explicitly about icmp, but this seems relevant to IKEv1
too. For some background, as per RFC 4301/5996, icmp type is put in the
most significant 8 bits and icmp code is in the least significant 8 bits
of port field. Although Openswan does not have any configuration options
for icmp type/code values, it is possible to specify icmp type and code
using protoport option. For example, icmp echo request (type 8/code 0)
needs to be encoded as 0x0800 in the port field and can be specified
as left/rightprotoport=icmp/2048.

Now with NETKEY, icmp type and code need to be passed as source and
destination ports, respectively. Therefore, code in the attached patch
(openswan-icmp-processing.patch) code extracts upper 8 bits and lower
8 bits and puts into source and destination ports before passing to
NETKEY. I have put this explanation in the patch too for better clarity.

Signed-off-by: Paul Wouters <paul@xelerance.com>

(This needs a new test case for both KLIPS and NETKEY)