Fix inbound policy --addin, add --replacein
As far as I can tell the --addin option is to set the inbound policy
for an SA,  but,  we were not setting the flags appropriately.
Fix this up and also add a replacein option for completeness.

This code is only used when using eroute to configure tunnels rather than
letting pluto handle it.

Currently the manpage does not document --addin (and now --replacein)
and needs updating.