Routing cache corruption due to ip_select_ident
The comments in the code said that dst has to be set before calling
ip_select_ident, but dst wasn't set correctly.

It was corrupting a private value (rt->rt6i_nfheader_len) that would then
result in ip6_output calculating invalid mtu/packet sizing and rejecting the
transmission of a packet with EMSGSIZE.

Rework the dst settings and ip_select_ident code so that is all gets done
in the correct sequence.  This helps clean up the flowi code a little and
it is easier now to clean it up properly at some point w.r.t. IPv6.