* OCF: Give a hard #error in ipsec_ocf.h if we don't have CONFIG_KLIPS_OCF
I managed to compiel with it defined in linux/net/ipsec/defconfig but
without using it in my MODULE_DEF_INCLUDE that pointed to the file
packaging/linus/config-all.h. This resulted in some obscure error, instead
of a clear error that the config files did not match up properly.