* Changed a few *alloc() calls to alloc_bytes()
We were not tracking some allocations for LEAK_DETECTIVE. Seeing that
I did not find the proper free() calls, it might mean these are actual
leaks (unless pfreeany() was used).

One realloc() call in ssdep_linux.c was left, as I was not sure how
to do the realloc() call using our wrappers.