I cancel a patch of pluto that deletes port floating codes when Delete SA is received.
The former part of the condition never matches any objects, that is
there's no object with st_serialno is 0.  After all, it is equivalent to
"st->st_serialno == nfo->st->st_serialno".

Test environment:
Server 0 -+- Server 1
          |
          +- NAT Router -- Server 2
Server 0 is a target to test.
NAT router is set up with:
 iptables -t nat -I POSTROUTING -j MASQUERADE --random

I set up 3 connections: nonat, nattun0, and nattun1

Here's a scenario:
1. start all openswan on Server 0, 1, and 2
2. Server 1 initiate "nonat" connection.
3. Server 2 initiate "nattun0" and "nattun1" connections.
4. "conntrack -D -p udp" on NAT Router and clear NAT table.
5. Server 2 delete "nattun1"
6. stop openswan on Seavrer 1, 2, and 0.

With the old code, "nonat" is rewritten when "nattun0" is established.
It's the issue that we wanted to fix.

With the current code, above issue is fixed, but "nattun0" is not
written when "nattun1" is deleted and port floating occurs.  Thus,
Server 0 can't send Delete SA to Server 2 when openswan on Server 2 stops.

With the patched code, both issues are fixed.

Signed-off-by: Paul Wouters <paul@xelerance.com>