* vhost allows connections with subnets proposed and ignores virtual_private
In virtual.c the function returned NULL at 364 if vhost is used and the
proposed net is not a host.

The bad thing is that at this point except for the fact that the net
itself is not allowed no further checks will be done and the proposed
subnet is accepted regardless if the proposed net is a public network,
the world (0.0.0.0) or whatever.

vhost:%no,%priv and virtual_private= are useless in this case.