* Fix for CVE-2011-4073 crypto helper crash
When a helper or helper queue had work for a phase2, and the
corresponding phase1 was deleted, the helper would not get informed.
Once it had completed its work, it would try and use the old pointers
to write the work item data.

This would only happen when not running with nhelpers=0. See the CVE
announcement for more details.