* IKEv2: Fix crash on max number of retransmissions reached STATE_PARENT_I2
In ipsecdoi_replace() there is a check for the kind of state we are in to
see if we should save some partial policy we got out of the connection so
far into the state. It did not account properly for the STATE_PARENT_I2
and so it thought it had more policy then there really was, hitting a
passert.

This showed up as:

[root@oswtest1 openswan.git]# ipsec auto --up test
133 "test" #3: STATE_PARENT_I1: initiate
133 "test" #3: STATE_PARENT_I1: sent v2I1, expected v2R1
134 "test" #4: STATE_PARENT_I2: sent v2I2, expected v2R2 {auth=IKEv2 cipher=aes_128 integ=sha1_96 prf=oakley_sha group=modp2048}
010 "test" #4: STATE_PARENT_I2: retransmission; will wait 20s for response
010 "test" #4: STATE_PARENT_I2: retransmission; will wait 40s for response
[...]
010 "test" #4: STATE_PARENT_I2: retransmission; will wait 40s for response
031 "test" #4: max number of retransmissions (20) reached STATE_PARENT_I2.  Possible authentication failure: no acceptable response to our first encrypted message
000 "test" #4: starting keying attempt 1 of an unlimited number, but releasing whack

At this point it would hit the passert.