* IKEv2: If responder narrows proposal, update Traffic Selector payload.
See http://tools.ietf.org/html/rfc5996#section-2.9

   When the responder chooses a subset of the traffic proposed by the
   initiator, it narrows the Traffic Selectors to some subset of the
   initiator's proposal (provided the set does not become the null set).

The function ikev2_port_in_range() was added to deal with the strange
situation of port ranges. We can only specify one number, which means
"all ports". Port ranges of 0-65535 and 0-0 both mean "all ports".

In ikev2_calc_emit_ts() when we are RESPONDER, we narrow the traffic
selectors with protocol and port ranges. Matching on more-specific
subnet ranges was already done by picking the best connection.

Note that we currently have no way of specifying a non-all range of
ports via our configuration. We log a warning if we see these in
the traffic selector,